Refactor: Standardize JS filenames and remove obsolete function
- I standardized JavaScript filenames to use snake_case.
  - I renamed `Anti Cheats BP/scripts/assets/globalBanList.js` to
    `Anti Cheats BP/scripts/assets/global_ban_list.js`.
  - I updated import statements in `player_event_handlers.js`,
    `initialize.js`, and a comment in `index.js` to reflect
    this change.

- I removed the obsolete function `Anti Cheats BP/functions/assets/asset.mcfunction`
  after confirming its logic had been migrated to JavaScript
  (in `scripts/systems/periodic_checks.js`).
