--- Anti Cheats BP/functions/assets/asset.mcfunction
@@ -1,4 +1,0 @@
-# This function is intentionally left empty.
-# The vanish announcement logic previously here (action bar message) has been
-# migrated to JavaScript in scripts/systems/periodic_checks.js
-# in the "Vanish Reminder Interval" for better control and to address the TODO.
