+++ Anti Cheats BP/scripts/assets/global_ban_list.js
